refactored into lanyard component
